Add GraphQL Playground 1.7.23
This is a POC from the 2020 codefest to provide GraphQL Playground as an alternative to GraphiQL.

GraphQL is built in React and provided as a React component. To ease the integration with this module, an embedded version of React has been used (see https://github.com/prisma-labs/graphql-playground/tree/master/packages/graphql-playground-html).

GraphQL Playground is built on top of GraphiQL (auto-completion, syntax highlighting, documentation, history/favorites) and provides some nice to have additions:
- Tabs!
- Improved documentation navigator
- Convenient GraphQL Schema
- Possibility to share queries (using https://www.graphqlbin.com/)
- Built-in "dark" and light" themes
- Customizable color scheme

It has been recently announced that GraphQL Playground will join forces with GraphiQL (see https://foundation.graphql.org/news/2020/04/03/web-based-graphql-ides-for-the-win-how-why-playground-graphiql-are-joining-forces/), and as a consequence it's likely that GraphQL Playground will no longer benefit from major updates. Integration target is GraphiQL 2 which is an on-going effort though.

As a side note, I find the integration of history/favorites much less convenient than the one provided with GraphiQL; no way to "rename" queries, history starting from least recent item, no history limit (stored in local browser storage, is it even purged at some point eventually?).